adding fov select
